lastchangeraed

+++ src/main/java/com/pi/farmease/dao/CartRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface CartRepository extends JpaRepository<Cart,Long> {
-    Cart findByUserId(Long userId);
+    Cart findByUserId(Integer userId);
 
 }
